feature/di-java

+++ 04_serlvets/servlets/src/main/java/ru/netology/service/PostService.java
@@ -1,32 +1,39 @@
 package ru.netology.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 import ru.netology.exception.NotFoundException;
 import ru.netology.model.Post;
 import ru.netology.repository.PostRepository;
 
 import java.util.List;
+import java.util.Optional;
 
+@Component
 public class PostService {
-  private final PostRepository repository;
+    private final PostRepository repository;
 
-  public PostService(PostRepository repository) {
-    this.repository = repository;
-  }
+    @Autowired
+    public PostService(PostRepository repository) {
+        this.repository = repository;
+    }
 
-  public List<Post> all() {
-    return repository.all();
-  }
+    public List<Post> all() {
+        return repository.all();
+    }
 
-  public Post getById(long id) {
-    return repository.getById(id).orElseThrow(NotFoundException::new);
-  }
+    public Optional<Post> getById(long id) {
+        return repository.getById(id);
+    }
 
-  public Post save(Post post) {
-    return repository.save(post);
-  }
+    public Post save(Post post) {
+        return repository.save(post);
+    }
 
-  public void removeById(long id) {
-    repository.removeById(id);
-  }
+    public void removeById(long id) {
+        if (!repository.getById(id).isPresent()) {
+            throw new NotFoundException("Post with id " + id + " not found");
+        }
+        repository.removeById(id);
+    }
 }
-
